Fix main directory compiler warnings. (#41)

+++ src/main/java/tfw/immutable/ila/byteila/ByteIlaDecimate.java
@@ -44,7 +44,7 @@ public final class ByteIlaDecimate {
             final ByteIlaIterator fi = new ByteIlaIterator(segment, bufferSize);
 
             for (int ii = offset; length > 0; ii += stride, --length) {
-                array[ii] = (byte) fi.next();
+                array[ii] = fi.next();
                 fi.skip(factor - 1);
             }
         }
