+++ src/main/java/tfw/immutable/ila/longila/LongIlaNegate.java
@@ -32,7 +32,7 @@ public final class LongIlaNegate {
             ila.toArray(array, offset, stride, start, length);
 
             for (int ii = offset; length > 0; ii += stride, --length) {
-                array[ii] = (long) -array[ii];
+                array[ii] = -array[ii];
             }
         }
     }
